Reduced size of the landing page logo to reduce overflow
If it's too big, there will be overflow and that's not really needed. The landing page should be rather minimal so as to not confuse new users; more information can always be provided in the actual documentation.

I THINK this will look a bit tiny on small-screen mobile devices, but the navigation breaks anyway if using docusaurus' default style there, so there likely needs to be some responsive design (media queries) added later that can resolve both issues.