Update landing page layout to eliminate scrolling
Obviously this requires a certain screen size/desktop layouts, but it makes the landing page much easier to survey at a glance. This should help especially new users by avoiding information overload.

Compare the node.js website, which has a landing page that is even more barebones. In the case of Luvit, some key information does need to be provided (I think) since few people will have heard of it, or even Lua, so I decided on this trade-off instead of culling all the text or moving it to another page.

On mobile/small screen sizes, the layout mostly adapts, though there are some responsive design issues that would still have to be sorted out (not difficult, but I haven't gotten around to it yet).